Revert to C dispatcher
While it's proven that the Rust dispatcher works, and that attaching
both C and Rust programs to functions of that dispatcher also works,
there are enough edge cases in the generated BTF from Rust programs
for this not to be considered stable enough for now.

Reverting to C, but will keep the Rust code here so we can switch out
once BTF support for Rust in LLVM/Rustc/bpf-linker/kernel is stable.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>