TEMP: Try and fix stale Maps + Dis Progs
Currenlty even when there are no programs
to load the dispatcher program is left
orphaned.  This takes a stab at ensuring
we remove it when there's no more programs
left to be managed.

This was in an effort to figure out why BPFmaps
were orphaned when a program was loaded and
promptly unloaded.

Signed-off-by: Andrew Stoycos <astoycos@redhat.com>