Fix off by 1 error
PR #54 fixed the issue when the first program is loaded, but there is
still an off by one. When `next_available_id` is calculated, it is using
`prog.len()` when at least one is already loaded. But at the time, the
program being inserted hasn't been inserted yet, so `+1` is needed.

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>