bpfd: Split add_program
This splits add_program so it can be used for any program type.
As such, the Program enum has become part of the internal API used to
communicate between the RPC thread and the BPF thread. This required
moving the ifname_to_ifindex and ifindex_to_ifname utility functions
into their own module as they are required in a few places.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>