proceed_on is always empty
If proceed-on is not set in the gRPC message, it will not be displayed
on the list (data is empty in the gRPC request). It is applied properly
because when writing to the dispatcher, it is converted to a bitmask
and a default mask is used for an empty vector.

Resolves: #157

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>