Make EbpfProgram and EbpfProgramConfig cluster scoped
Make these two objects cluster scoped since the concept
of an ebpf program running on a node cannot be
constrained to a namespace.  I.E a program running
on a single node could possibly affect all pod traffic
on that node reguardless of namespace.

TODO, we might need to be clever if we want to express
selecting something like: "Attatch a given TC program
to all host bound interfaces of pods in ns A."

Signed-off-by: Andrew Stoycos <astoycos@redhat.com>