Remove passing config to example counters via toml
Passing config to the examples (go-xdp-counter and go-tc-counter) using a toml
file was used in k8s deployments prior to the bpfd-operator work. With the
new design, there is no need to pass config to the userspace container. It
now reads the bpfprogram CRD to ge the UUID.

Reading of the toml file is still in place for running locally so the cert file
locations can be overwritten.

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>