Setup prod logging, fix races
This commit sets up and allows the user to configure logging for
both the bpfd-operator and bpfd-agent via the `RUST_LOG` env var.

It also fixes a few bugs in the operator related to failing to
update stale resources, over-reconciling (getting extra reconcile signals),
and a small edge case related to writing an incorrect "successful" status
at startup.

In the agent we make sure to write the correct status after successfully
loading a program.

Signed-off-by: Andrew Stoycos <astoycos@redhat.com>