Additional proceedOn fix
PR #298 moved around where the loadRequest message was being built. The issue
is that the loadRequest message was being built after some logic to clear out
the proceedOn fields (TC and XDP have different defaults and it is difficult
to do a proper deepEqual with the value set). So the loadRequest was being built
with a cleared out ProceedOn value. Moved the building of the loadRequest to
before the logic to clear the fields.

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>