Build shared client library
Provide the foundation for a shared library which applications
can use helper functions from i.e
to get maps from a bpfProgram kubernetes object.

The library can also be used to for auto-discovery if
an application is trying to decide wether or not to load
via bpfd.

Lastly this library can be used to programmatically create and
manage bpfProgramObjects.

Add Wait function

Add ways to delete a bpProgramConfig

Add helper to detect if bpfd is installed on a system.

Add create/update owned helper

Some operators may want to create a bpfProgramObject as an
owned object to ensure automatic cleanup.

Add a helper to allow this.

Signed-off-by: Andrew Stoycos <astoycos@redhat.com>