Update Multi-attach definition to use InterfaceSelector
The K8s CRDs (BpfProgramConfig and BpfProgram) use an `Interface` field in the
`BpfNetworkMultiAttach` structure to indicate which interface the BPF program
for XDP or TC should be loaded on. Going forward, need a better was to selector
interface, like a pod label or the main interface on a node. This PR pushes
the interface down a level to allow for more options in determining the interface
to load the BPF program on.

Modify BpfProgramConfig to allow the Primary Interface on a Node to be
specified as the interface. This is special in the fact that the
BpfProgramConfig (which is a flag indicating primary interface) does not
match the value on BpfProgram, which is the interface that the flag maps
too.

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>