Add ImagePullPolicy to the bpfd api
Add the ability for a user to specify an `ImagePullPolicy`
for the bpfd-api.  This policy controls when a bytecode image is
pulled from a remote repository, it's either `Always`, `IfNotPresent`,
or `Never`.

Add plain auth capability to the api as well to allow users to authenticate
with private image repositories.

Lastly cleanup some of the api type helper functions to be more consistent.

Signed-off-by: Andrew Stoycos <astoycos@redhat.com>