Add new oci-utils mod fix static_program mod
With the addition of the ability to specify container registry credentials
and image-pull-policy this creates a new module, oci-utils, to handle all the
new functionality.

This new module creates a custom image content store, which ensures
the validity of data on disk before loading AND makes sure we can adequately
store container image on the local host.

This also comes with a cleanup of API types, specifically we try to standardize
how conversions between internal and external api types are managed.

Finally this fixes the static_programs functionality to make sure it can support
all program types, and is a bit more isolated in functionality to it's own module.

Signed-off-by: Andrew Stoycos <astoycos@redhat.com>