Pin tracepoint links
This PR allows bpfd to pin tracepoint program
links in order to allow the program to keep running
even after the loading thread has exited.

It also improves the error reporting for failure to pin
links vs pinning programs in bpfd as well as for when
a tracepoint program cannot be attached.

Point to a custom aya branch for now to pick up these
changes.

Signed-off-by: Andrew Stoycos <astoycos@redhat.com>