bpfctl should use TCP socket over TLS if Unix Socket not available
Currently, if the unix socket setup fails, the `bpfctl` command stops executing
and returns an error. The code should try the TCP socket if the Unix socket
fails.

```
$ bpfctl list
Error: transport error

Caused by:
    0: error trying to connect: Permission denied (os error 13)
    1: Permission denied (os error 13)
```

FYI: The reason for the failure was that the user didn't have write access
to the socket. It works if `sudo` is used. This is fixed in future commit.

Also added a log to indicate which method, unix or TCP, was being used:
```
$ sudo RUST_LOG=info bpfctl list
[2023-04-21T15:28:23Z WARN  bpfd_api::config] No config file provided. Using default
[2023-04-21T15:28:23Z INFO  bpfctl] Using UNIX socket as transport
 UUID  Type  Name  Location  Metadata

$ RUST_LOG=info bpfctl list
[2023-04-21T18:06:06Z WARN  bpfd_api::config] No config file provided. Using default
[2023-04-21T18:06:06Z INFO  bpfctl] Using TLS over TCP socket as transport
 UUID  Type  Name  Location  Metadata
```

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>