Finish implementing support for TC proceed-on (issue #299)
A key point to understand about this implementation the following:

Valid TC return values range from -1 to 8.  Since -1 is not a valid shift value,
1 is added to the value to determine the bit to set in the bitmask and,
correspondingly, The TC dispatcher adds 1 to the return value from the BPF program
before it compares it to the configured bit mask.

I also tweaked the display for bpfctl list a little.

Signed-off-by: Andre Fredette <afredette@redhat.com>