Bump tonic to 0.9.2 and rustls to 0.21.1
The tonic bump to 0.9.2 brought in rustls 0.20.0. With this bump, as well as
the bump of OpenSSL to 0.10.52, mTLS between client and server were broken.

There was a bug in rustls (see https://github.com/rustls/rustls/issues/1288)
where the certificate created by the rust server could not be parsed by the
GO client (but an identical OpenSSL based certificate could be parsed). This
was fixed in rustls 0.21.1.

rust-openssl was also updated in this timeframe, which cause an error parsing
the rust generated certificate in the rust client (bpfctl). See
https://github.com/rustls/rustls/issues/1292. This is fixed by this PR by
changing how the certicate was created. The `.dns()` line below was causing the
problem. A partial fix was implemented in #448, which fixed the GO clients,
but the rust clients were still broken.
```
    let subject_alt_name = SubjectAlternativeName::new()
        .dns("localhost, IP:127.0.0.1")
        .build(&cert_builder.x509v3_context(Some(&ca_cert_x590), None))?;
```
As part of this fix, fixed up the `scripts/certificate.sh` file to generate
the certificates properly using OpenSSL. This helped a lot in debugging this
issue. Also rearranged the `bpfd/src/cert.rs` file to match the certificate
created in the script so they lined up side-by-side.

Fixes: #398
Fixes: #411
Fixes: #424

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>