bpfctl: wrap Table into own ProgTable type
List command uses comfy_table::table::Table to collect and show the
results. Some of the code is application specific and can be hidden
from the client code.

Use own type with own constructor to hide table structure and
provide add_response_prog API which hides all the handling of
response from the caller code.

Add add_row API with row data as separate parameters. In addition to
moving vector creation to the common place it provides additional
clarity and check for the row data.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>