Fix tracepoint example when run on host
Tracepoint example run from host doesn't work because it has wrong sectionname:
```
$ ./go-tracepoint-counter
2023/05/02 12:02:45 Using Input: Interface= Priority=50 Source=0xc0005fd8f0
2023/05/02 12:02:45 Reading /etc/bpfd/gocounter.toml ...
2023/05/02 12:02:45 Read /etc/bpfd/gocounter.toml failed: err open /etc/bpfd/gocounter.toml: no such file or directory
2023/05/02 12:02:45 failed to load BPF program: rpc error: code = Aborted desc = Unable to find a valid program with section name stats
```
Change sectionname from "stats" to "tracepoint_kill_recorder".

When stopped using <Ctrl-C>, program is not unloaded. Rework code to look for the
os signals.

Replace all log.Fatalf() calls with a log.Printf() and a return. When log.Fatalf() is
used, all defers are no longer called, so things may not cleanup properly.

Nit: Also changed the output string such that the count is at the end of the output.
The counts blended in with the time/date stamp and made it hard to read at a
glance.
Changed from:
```
2023/05/02 12:35:23 5 SIGUSR1 signals seen
2023/05/02 12:35:24 8 SIGUSR1 signals seen
2023/05/02 12:35:25 12 SIGUSR1 signals seen
```
to
```
2023/05/02 12:35:23 SIGUSR1 signal count: 5
2023/05/02 12:35:24 SIGUSR1 signal count: 8
2023/05/02 12:35:25 SIGUSR1 signal count: 12
```

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>

Squash

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>