Cleanup examples helptext
Added newlines to the helptext on the examples to make it easier to read.
Also removed the unused variables from the tracepoint example (interface
and priority) helptext.

There was also an additional parameter (kubeconfig) being pulled in from
container-runtime (via bpfd-helpers). Tried to remove bpfd-helpers from
param.go, but that didn't fix. But left the change in to remove an extra
dependency in param.go. Adding `flag.NewFlagSet()` removed it.

I assume it was part of this change, but the tc kubernetes integration
test starting failing. The test code was looking for "sent" in the output.
Not sure how it was working before, but it is running the
examples/go-tc-counter/kubernetes-deployment/go-tc-counter-bytecode.yaml
which is running with direction of ingress. And ingress prints "received".

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>