Improve endpoint configuration
Endpoint configuration now allows multiple unix and TCP sockets to be
used for use as both server in bpfd and client in bpfctl. Individual
endpoints can also be enabled and disabled via configuration. For client
connections, priority is determined by the order the endpoints are
listed in the TOML configuration file.

Fixes #418

Signed-off-by: Mauro Ezequiel Moltrasio <mmoltras@redhat.com>