Updates to API files
I noticed recently that bpfd-api/src/bpfd.v1.rs was getting changed when
I ran cargo xtask build-proto.  I'm not sure what caused the change, but
when that happened, the new file didn't get checked in.

Just to check, I also updated rust and did a dnf update and tried again.
The dnf update pulled in a new version of protoc.  This resulted in
updates to the go api files, but the only change is to a comment listing
the new version of protoc.

Signed-off-by: Andre Fredette <afredette@redhat.com>