copy examples bytecode to /run/bpfd/examples
Recent changes caused all file open calls to be run on one of the spawned
threads (without caps) instead of the main bpfd thread. As a result, when
bpfd is run as a systemd service, the spawn thread can't access files not
owned by bpfd.

The causes an issue with the example code because the bytecode (bpf_bpfel.o)
was being loaded directly from the github branch, which was owned by the
developer, not bpfd user.

Updated the install scripts to also copy the example bytecode files into a
bpfd owned directory (/run/bpfd/examples/). Also updated all the examples
to load from this new location if no source is provided. For when bpfd is
not run as a systemd service (setup.sh install not called) and the bytecode
is not copied, also check for a local copy (like was done originally).

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>

Squash

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>