bpfctl: LoadCommands: factor out main functionality
Both LoadFromImage and LoadFromFile share quite a lot of common
functionality to define LoadRequestCommon structure. Move it in one
place as a method for Commands type (because it needs to access
parts of LoadFromFile and LoadFromImage variants but LoadFileArgs
and LoadImageArgs are different types). Comparing to prog_type and
attach_type it's not exactly the same since location is specific for
particular command.

It could make sense to make LoadRequestCommon inside the match
branches directly, but in this particular case there is some
processing of the data which is better to keep in one
place (id.clone(), section_name.to_string() and so on) and the code
extracts `global` field from inside of different enum types. So,
initialize the common values before LoadRequestCommon creation.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>