Change the way we save TC links
The behavior for TC links changed when detaching links on drop was
fixed in Aya. As a result, TC links started getting dropped when bpfd
would exit.

With this change, we don't save the link in TcDispatcher anymore.
Instead, we mem::forget the link and then recreate it when it needs
to be detached.  As a result, the link doesn't get dropped when bpfd
exits. This also fixes the reload issue.

Fixes: #501

Signed-off-by: Andre Fredette <afredette@redhat.com>