handle cases where the old dispatcher isn't attached
If the old dispatcher was not attached when the new dispatcher
gets attached, the new dispatcher may get the same handle
that the old one had. If this happens, the new and old dispatchers
will have the same parameters, so we'll inadvertently detach the
new dispatcher when we attempt to detach the old one. This puts us
in a vicious cycle in which the new dispatcher gets attached and
then immediately detached.

Also, handle the error when attempting to detach a dispatcher that is
not attached.  In this case we can just ignore the error, but it is an
unexpected event, so I'm logging a debug message.

Signed-off-by: Andre Fredette <afredette@redhat.com>