Switch to a single BpfProgram Object per bpfd prog
Ensure a 1 to 1 mapping between bpfd programs on a node
and bpfProgram K8s Objects.

To make this happen many core changes needed to occur such as adding
new methods to the bpfd-agent common interface and adding some
new smarter bpfProgram predicate functions to the controllers.

Additionally update some constants and fixup examples, helpers, and tests
for this change.

Signed-off-by: Andrew Stoycos <astoycos@redhat.com>