Add e2e testing to bpfd integration test
The existing integration tests validate the basic loading, unloading,
and cli functionality for bpfd.  However, they do not validate that the
programs are actually running or that any of the advanced features
actually work.  This commit adds support for these functions to include
the following.

eBPF Integration Test Framework:
- Create and use a network namespace with a veth pair and IP addresses
  for testing XDP and TC programs
- Start, capture, and retrieve the output of pings to the network
  namespace to generate traffic for these programs
- Capture and retrieve /sys/kernel/debug/tracing/trace_pipe so that
  bpf_printk() logs from eBPF programs can be examined

The above provides a framework with which functionality can be easily
tested by adding and examining bpf_printk() output to/from the eBPF
programs being installed to confirm that they are working correctly.
Using the namespace has the added benefit that we can add XDP and TC
programs to the veth that we created and not bother the user with having
to configure an interface to use.

Additionally, tests have been added for the following functionality
using this framework:
- the programs loaded are actually running
- setting of global variables
- priority for XDP and TC programs
- proceed-on for XDP and TC programs
- tests for TC egress in addition to ingress

Fixes: #394

Signed-off-by: Andre Fredette <afredette@redhat.com>