build example us images in CI
Build the example's userspace images in actions prior to running
the kubernetes integration tests so that they can be loaded into
the kind cluster beforehand. This ensures that the tests are always
run with up-to-date userspace example images.

Also fix a small todo which cleaned up the bpfd deployment following a
test run to ensure the deletion process functions correctly.

Signed-off-by: Andrew Stoycos <astoycos@redhat.com>