allow multiple attachments in a single *Program
This commit includes some refactoring changes which allows the user to
load a bpf program to multiple attachment points in a single *Program
kubernetes object. Specifically it makes it much easier to attach the same
XDP/TC program to multiple interfaces OR a single TracePoint Program to
multiple tracepoints.

It also adds some unit tests to cover these cases.

Signed-off-by: Andrew Stoycos <astoycos@redhat.com>