bpfd-operator: Reduce the number of unload/reloads
In bpfd-operator, the compare function DoesProgExist() compares the current
LoadRequest values to the what is loaded in bpfd in the ListResult. The compare
was comparing the pointer values of the Id instead of the actual strings, so
it always indicated that there change. This caused the calling function to
always unload and reload the EBPF Program.

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>