Remove check to block unloading of a Map Owner
There was a check in bpfd which prevent an eBPF program from being deleted
if it's maps were being referenced by another eBPF Program. The check was
added during the original map sharing PR because it caused problems, but
that is no longer the case.

bpfd used to remove the Map Owner from the list of UUIDs in the MapsUsedBy
field, because it looked cleaner. The Map Owner is now included in the list
of UUIDs in the MapsUsedBy because the Map Owner eBPF Program can now be
unloaded, so it is useful to know if it is there or not.

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>