Handle and report bytecode selector errors
With the original bug, if there was a problem with GetBytecode(), the
bpfProgram object would have been created, but the program would fail to
load because GetBytecode() would fail and return.  We didn’t tell the
user why it failed and you couldn’t delete the *program object because
in order to do that, you need to call reconcileBpfdProgram(), which
requires the bytecode interface.

Andrew’s recent pr changed the behavior so that the bpfProgram objects
get created only after the GetBytecode() succeeds.  This is better
because you don’t have the bpfProgram object that can’t be deleted, but
because you don’t have the bpfProgram object, you don’t know why it
failed unless you look at the agent logs.

Both versions still had a problem where if the secret gets deleted (or
changed so it doesn’t work after a bpf program is loaded) you can’t
unload programs that use it even though the bytecode  interface isn’t
necessary for unloading programs.

This pr moves the GetBytecode() call into reconcileBpfdProgram(), and
only calls it when we need to do a load.  This allows the bpfProgram
object to get created and reports the error with a new condition called
BpfProgCondBytecodeSelectorError.  It also allows you to delete the
program.

Fixes: issue #640

Signed-off-by: Andre Fredette <afredette@redhat.com>