Modify updateStatus() to remove condition entry before readding
In the updateStatus() function for both bpfd-agent and bpfd-operator, modify
the function to remove an updated condition before readding it. The
meta.SetStatusCondition() does not add a condition if it is already in the
list and doesn't update the TimeStamp if Status hasn't changed. Plus, the
code expects the last entry in the list to be the lastest condition.

See: https://github.com/kubernetes/apimachinery/blob/v0.27.4/pkg/api/meta/conditions.go#L30

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>