bpfd: use prog id as the one true identifier
A bunch of internal changes in order to make the kernel id the one
true identifier internally in bpfd.

The majority of the changes were related to the fact that we no longer
know the id ahead of load time. Generally though 90% of the changes
were simply superficial and shouldn't affect any of the existing flow.

Signed-off-by: astoycos <astoycos@redhat.com>