Add support for Get API
bpfd does not support a Get API. To retrieve a given program, the application
must retrieve all programs and filter out the program it needs. This PR adds
a Get API to bpfd, which allows an application to retrieve a given program by
it's program id. As part of this PR, the Load commands also return the same
data returned by the Get for the program that was just loaded.

The API structures were also reworked so that all the Kernel Data is in its
own structure and the bpfd based data is in its own structure. These new
structs are returned by the List, Get and Load commands.

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>