Just keep one condition for each program object
All of the code assumes that only the last condition counts. However,
that's not how conditions work. This change modifies our behavior so
that we only maintain the one (most recent) condition for each program
object, which is a step toward the resolution of issue #430.

Part of the solution is to create a common updateCondition() function
for the bpfd-operator that is used for all program types which removes
some code duplication and is a small step toward the resolution of
issue #709.

This also removes code duplication in xdp-program_test.go, which is a
small step toward the resolution of issue #710.

Signed-off-by: Andre Fredette <afredette@redhat.com>