Fix Procceedon bug
The procceedon settings were incorrect for XDP and TC after unload and
for TC after bpfd restart.

The issue was that we were adding a program to the list before sorting
in set_program_positions() whether we were adding, deleleting or
rebuilding.  For the later two cases this caused the actual positions
to be shifted.

Fixes 791

Signed-off-by: Andre Fredette <afredette@redhat.com>