Relicense userspace to Apache 2.0 only
This is simply dropping requirement for inbound contributions to be MIT
licensed + removal of outbound option of MIT.

As all contributions are already submitted under both licenses this
shouldn't be controversial. The reasoning for this is to simply our
license arrangement before submitting to the CNCF sandbox.

At some point in future we might revisit this and look to provide MIT
as an outbound option again - which is very common in the Rust
ecosystem.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>