bpfd, bpfctl, operator: Remove support for TCP/TLS
This removes support for the use of TCP/TLS endpoints.

bpfd/bpfctl will default to using the unix socket.
Permissions there may be controlled using users/groups.

The bpfd-operator will also use the unix socket but through a volume
mount that is shared with bpfd.

TCP/TLS isn't required for any practical reason - we only want to
support on-the-same-box communication - and the dependencies required to
implement this support were vast. We may chose to revisit this in
future.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>