bpfman: add support for socket activation
When loading bpfman as a systemd service, implement socket activation.
Socket activation is a systemd feature in which a given service is not
started until it is needed, in this case, a request is received on a
socket. To reduce the attack surface, it is preferred that bpfman not
run forever, but instead stop after a period of inactivity. The current
implementation has a hard stop of bpfman after 15 seconds. Trying to
restart timers over multiple async threads was not worth the effort
since async nature of bpfman is going to be ripped out over the next
couple of weeks.

NOTE: Systemd can support stopping a service after a specified time
of inactivity. But it requires the builtin systemd proxy, where systemd
listens on the primary socket and when a request comes in the request
is sent to bpfman over a private socket. It was decided that this was
not the correct solution for bpfman.

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>