bpfman: in memory db setup + oci_utils conversion
Prep for using sled as the in memory db for all state persistence so that
we no longer have to write directly to disk.

Add some error handling fixes.

Convert to using sled for the OCI image store. This ensures that no
data is written to the filesystem directly and that everything goes
through our sled db instance.

Signed-off-by: Andrew Stoycos <astoycos@redhat.com>