integration-test: edit pull_bytecode test
Edit this test with the migration to using sled's in memory db because
sled doesn't currently support multiple "open" instances, so it's impossible
for use to verify the db state from external processes. Rely on our
unit test coverage for the underlying functionality and just use this
as a smoke test to ensure no panics occur.

Signed-off-by: Andrew Stoycos <astoycos@redhat.com>