bpfman: Remove grpc section from config
At one point, the ConfigFile (/etc/bpfman/bpfman.toml) could be used to configure
the gRPC endpoint as TCP/TLS or Unix socket. TCP/TLS was removed and now only Unix
Sockets are valid, so this configuration is now useless. So removing the gRPC
Endpoint from the ConfigFile.

bpfman-agent was only using this field from the ConfigFile, so it no longer needs
to read the ConfigFile. So all ConfigFile logic was removed from bpfman-operator.
The only gotcha is that bpfman-agent was using the gRPC Endpoint to change the
socket path from the default path of /run/bpfman/bpfman.sock to
/bpfman-sock/bpfman.sock. This was because the bpfman agent doesn't have
permissions to access the host. To work around this fact, the default path was
moved to a subdirectory of /run/bpfman/sock/bpfman.sock.

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>