operator: make health and metrics ports configurable
In some scenarios, the health and metric ports may are already in use by
other services on the system. When this happens the bpfman-agent container
fails to deploy. The current default is 8080 and 8081. First, move the
port numbers to something less common. Then make the values
configurable.

The ports are passed in through the yaml files. However, the bpfd-agent
releated yaml files are built into the bpfman-operator image, so
the values cannot be easily changed without rebuilding the image. So
add fields to the ConfigMap so they can be overwritten without
rebuilding the bpfman-operator image.

Resolves: #914

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>