bpfman: add retries to sled access
To support removing RPC from the bpfman CLI, add a retry to the sled DB
access so whien multiple simultaneous CLI calls are made, the additional
calls won't panic immediately but will retry after a delay.

The ROOT_DB is a `lazy_static`, which doesn't have a way to pass in
parameters. So switch to using `once_cell`. It is simliar in that it
doesn't initialize until the first time it is called. However, the
initializaton is in a separate function (which can be called multiple
times). So ROOT_DB still exists, but to get the value, use the `get()`
function, `ROOT_DB.get()`. It now returns an Option, so all ROOT_DB
instances need a `get().expect()`. A new initialization function was
created, root_db_init(), which takes `config` as a parameter (`config`
is the marshalled data from the `bpfman.toml` file).

Added a database `max_retries` and `millisec_delay` to `config` data in
the `bpfman.toml` file.

Cleaned up some `bpfman system service` helptext.

ISSUE: The `new_pre_load()` function in command.rs needs to call
`root_db_init(config)`, but `config` is not accessable from this
function at the moment and it would need to be passed though RPC to get
it there. The fix is probably to move that DB access to bpf_manager. A
short term fix may be to reread `bpfman.toml` here instead of trying to
pass it through.

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>