bpfman: remove RPC for CLI Get Command
As a first step to removing RPC from all the CLI commands, remove RPC
from the GET command and call bpf_manager Get function directly.

For the CLI Get code, removed the channel creation and RPC struct
building. Then create a bpf_manager instance and call
bpf_manager.get_program() directly. The bpf_manager call returns a
command::Program structure. The code is mapping from bpf_manager struct
to bpfman::v1 structs used in the original RPC calls, because thats
what the ProgTable expects. Once the ProgTable is reworked to print the
bpf_manager struct, this mapping can be removed.

The CLI system service code is what is used to launch bpfman as a
process. It contains the intialization code for bpfman. The
initialization code was broken into into 2 parts.
  - initialize_bpfman(): Does the directory creation and log
    initialization. Needs to be called by all CLI commands.
  - execute_service(): Now calls initialize_bpfman() but also calls
    serve(), which will spawn image_manager, etc, Only needed by
    SYSTEM/LOAD/UNLOAD/PUL. Not needed by GET/LIST

In the instantiation of `bpf_manager`, made `commands` and
`image_manager` Options, so that Get and List does not need to do the
extra work of spawning `image_manager` thread when it won't be used. As
part of this change, need to check that they are valid before using.

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>