Fix broken links in docs
The main issue reported was caused by the fact that the
"Setup and Building bpfman" link in README.md didn't work.

If you just try to go to bpfman.io, it sends you to the "main" version.
However, if you use bpfman.io with the path to a specific page without a
version in the URL (e.g.
https://bpfman.io/getting-started/building-bpfman), the link defaults to
v0.3.1, and the building-bpfman page has moved since then.

The fix employed was to add "/main" after "bpfman.io" in the links to go
to the current version.

There was also an https://bpfman.netlify.app link in CONTRIBUTING.md.

This commit also fixes a few other links, and in files that were
changed, the md formatter also made some formatting changes.

Signed-off-by: Andre Fredette <afredette@redhat.com>