remove channels init image_manager selectively
Remove all of the channel functionality
that was once used to receive commands, and instead
call the image_manager directly.

Push the initialization of the image manger into the
core codepath (bpf.rs) so that it can be selectively
initialized and used ONLY when it's needed.

In all places where the image_manage is expected
to be initialized panic with context if it's not
so that errors can be quickly discovered at runtime.

Signed-off-by: Andrew Stoycos <astoycos@redhat.com>