examples: Add uprobe example program
Other changes:

This pr includes a "go-target" program and k8s deployment that is
used as a target for the uprobe example on k8s.

The kustomize directories inadvertantly added for the kprobe example
are removed.

I also refactored the Makefile so that the instructions that were
repeated for each program could be shared by common deploy_prog and
undeploy_prog targets.

Fixes: #1021

Signed-off-by: Andre Fredette <afredette@redhat.com>

Makefile changes

Signed-off-by: Andre Fredette <afredette@redhat.com>