turn bpfman into a rust library
This commit involves a bunch of re-organization.
It mostly moves the cli portion of the codebase
to be it's own binary, and turns the core
functionality into a library both of which
live within the bpfman package.

This mainly involved making a bunch of functions
public and moving bits around.

Signed-off-by: Andrew Stoycos <astoycos@redhat.com>