chore(bpfman): Use native-tls
In order to be more easily packaged by Linux distros, we'll opt to use
native-tls over rust-tls. This puts a compile-time dependency on having
openssl libraries/headers available. This requires changes to the
Dockerfile and Specfile.

This also drops support for musl libc - given that figuring out how
to cross-compile OpenSSL for musl is a pain.

Therefore, we move away from using a `scratch` image to a RHEL UBI
image, which does include libc.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>