setup gopath in gh actions
The k8s controller generation code is really awful
and cannot output code to a custom filepath, therefore
when running `make generate` during the
generation verification process in actions
it dumps the code in `./github.com/bpfman/bpfman`
(i.e the package name) and causes builds to then
fail after the fact.

This makes sure that the code is checked out in
GOPATH/src/<PKG_NAME> which will allow the generated
code to be put in the right place.

Signed-off-by: Andrew Stoycos <astoycos@redhat.com>