bpfman: dir bpfman-sock missing for bpfman-rpc
When running bpfman as a long running process, bpfman-rpc fails to run
because directory /run/bpfman-sock is not created.

```
$ sudo RUST_LOG=info ./target/debug/bpfman-rpc --timeout=0
Error: No such file or directory (os error 2)
```

Second issue is that no logs are generated when bpfman-rpc is run.

As part of the rework of bpfman recently, the creation of
/run/bpfman-sock was moved from bpfman to bpfman-rpc, but it was
inadvertantly moved to a CSI only check. The fix is to move it out of
the CSI only logic.

Now that bpfman core is a library, all log initialization should occur
in the calling binary. So move the log initialization to bpfman-rpc and
bpfman CLI.

Resolves: #1077

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>