WIP: Add a node status field to *Programs
The NodeStatus field for each *Program is updated by the bpfman
controllers on each node.  With this change, the Operator
controller doesn't have to watch all bpfPrograms, but only the
NodeStatus field in the *Programs. This also distributes some of the
processing that the controller currently has to do across all the nodes,
so it should scale better.

I implemented two options:
1. NodeStatus []NodeStatusEntry
2. NodeConditions []metav1.Condition

I think the NodeConditions option is a twisted use of conditions and
the NodeStatus option is cleaner. So, my preferfence is NodeStatus.

This only includes the bpfman agent part of updating the
NodeStatus/Conditions. The controller would need to watch the node level
status and determine a global status for each *Program.

TODOs:
1. Investigagte using Server-Side Apply.
2. Implement operator controller support.
3. Scale testing.
4. Fix samples/bpfman.io_v1alpha1_kprobe_kprobeprogram.yaml (I changed
   it so I had a failing test case).

Signed-off-by: Andre Fredette <afredette@redhat.com>