fixup the rpm spec license
Although the bpfman project includes three
licenses, Apache-2.0, GPL2 and BSD2 the latter two
are there in order to support all of the eBPF
programs which we maintain in the xdp/tc dispatchers
as well as within the example programs.  Then
those eBPF programs are packaged and distributed
independently of the core code base via OCI container
images. The core bpfman binaries, bpfman, bpfman-ns, and
bpfman-rpc make up the bpfman rpm are all solely
licensed as Apache-2.0, therefore the spec file
needs to be updated to reflect that.

Signed-off-by: Andrew Stoycos <astoycos@redhat.com>