Temporary work-around for multi-program container images
The BpfApplication CRD support is intended to allow an application to
define a set of BPF programs required for operation.  This set of
programs may included multiple types of BPF programs that are packaged
in a single .o and single bytecode image file.  However, our currentl
image spec is designed to support a single bpf program with a single type.

More importantly, the bpfman code has a check to validate that the name
of program being loaded matches the name that is provided by the user.

As a temporary work-around, this pr just logs the mismatch and doesn't
fail the load.

A more complete solution will be provided by pr #1141

Signed-off-by: Andre Fredette <afredette@redhat.com>