docs: convert notes
Currently, when something is noted, something like `> ***Note:*** ...`
is used, which bold the word `Note:` and shifts the text over with |
displayed. We already have a markdown_extension that will display a box
around the note, so convert all the notes to use `!!! Note` to improve
the display of notes.

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>