Don't use S390 object code for generating build args
For generating build args from a cilium directory, we get the list
of bpf object files in the directory and then use the first one in
the list to generate the build program and map labels.  However,
Aya Object::parse doesn't currently work for S390 files, so if an
S390 file happens to be the first one in the list, it doesn't work.

This is a temporary fix that doesn't use S390 files, and if no
other type of file exists, it returns an error.

Fixes: #1200

Signed-off-by: Andre Fredette <afredette@redhat.com>