bpfman: Failed dispatcher load corrupts DB
If the TC or XDP dispatcher fails to load, subsequent load commands fail
with a panic in bpfman. To fix, removed an unwrap(), add a dispatcher
delete in the Dispatcher.new() function when an error occurs, and mapped
some errors to new BpfmanError for dispatcher failures instead of just
returning lower level error.

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>