bpfman: Make whether to use cosign configurable
This pr adds an option to Config called verify_enabled that controls
whether bpfman checks the signatures on images.

verify_enabled is currently defaulted to false, due to the current issue
with TUF (https://github.com/bpfman/bpfman/issues/1241).  It should be
set defaulted to true after the issue is resolved.

The handling of reading the config and setting defaults is cleaned up as well.

Fixes: #1239

Signed-off-by: Andre Fredette <afredette@redhat.com>