ci: run image-build.yaml when Conterfiles are updated
Currently, image-build.yaml, which is used to build and push all bpfman
related container images to quay.io, is run when PR is merged or a PR is
updated and image-build.yaml file itself is modified. A new rule needs
to be added such that image-build.yaml is run when a PR is updated and
one of the Containerfile files is modified.

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>