ci: bpfman image build needs same build and base image
bpfman images are built in github actions. The binary is built using
.github/wokflows/image-build.yml which is currently using ubuntu:latest.
Then the image is build by copoying the binary into a container. The
base image for the container is redhat/ubi9-minimal, which used glibc
2.34. Recently, ubuntu:latest change from ubuntu 22.04 (glibc 2.34) to
ubuntu 24.04 (glibc 2.39). So now bpfman won't run in a
redhat/ubi9-minimal base image.

This PR moves the build image back to ubuntu 22.04 (glibc 2.34). A
follow-up PR #1253 will change how this works and use the same build and
base images.

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>